refactor: move utility enums into feature-owned packages
