+++ src/main/java/com/codenames/codenames/backend/chat/ChatDto.java
@@ -1,6 +1,6 @@
 package com.codenames.codenames.backend.chat;
 
-import com.codenames.codenames.backend.utility.ChatMessageType;
+import com.codenames.codenames.backend.chat.api.dto.ChatMessageType;
 
 /**
  * Record class for messages that is used in controller.
