+++ src/main/java/com/codenames/codenames/backend/game/domain/Card.java
@@ -1,6 +1,6 @@
 package com.codenames.codenames.backend.game.domain;
 
-import com.codenames.codenames.backend.utility.Color;
+import com.codenames.codenames.backend.game.domain.Color;
 import lombok.Getter;
 
 /** Represents a single card on a board. */
