+++ src/main/java/com/codenames/codenames/backend/lobby/Lobby.java
@@ -1,8 +1,8 @@
 package com.codenames.codenames.backend.lobby;
 
 import com.codenames.codenames.backend.lobby.domain.Player;
-import com.codenames.codenames.backend.utility.Role;
-import com.codenames.codenames.backend.utility.Team;
+import com.codenames.codenames.backend.lobby.domain.Role;
+import com.codenames.codenames.backend.lobby.domain.Team;
 import java.security.SecureRandom;
 import java.util.HashMap;
 import java.util.List;
