+++ src/main/java/com/codenames/codenames/backend/lobby/domain/Role.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames.backend.utility;
+package com.codenames.codenames.backend.lobby.domain;
 
 /**
  * Represents the role of a player in the game.
